[AMS-4] domain - 자산 유형 수정, 삭제 기능 추가

+++ domain/src/main/java/com/sewon/assettype/application/AssetTypeService.java
@@ -1,7 +1,10 @@
 package com.sewon.assettype.application;
 
+import static com.sewon.assettype.exception.AssetTypeErrorCode.ASSET_TYPE_DELETE_ERROR;
+import static com.sewon.assettype.exception.AssetTypeErrorCode.ASSET_TYPE_DUPLICATION;
 import static com.sewon.assettype.exception.AssetTypeErrorCode.ASSET_TYPE_NOT_FOUND;
 
+import com.sewon.assettype.dto.AssetTypeParentResult;
 import com.sewon.assettype.model.AssetType;
 import com.sewon.assettype.repository.AssetTypeRepository;
 import com.sewon.common.exception.DomainException;
@@ -28,12 +31,31 @@ public class AssetTypeService {
         assetTypeRepository.save(assetType);
     }
 
+    @Transactional
+    public void updateAssetType(Long typeId, String type) {
+        if (assetTypeRepository.findByName(type).isPresent()) {
+            throw new DomainException(ASSET_TYPE_DUPLICATION);
+        }
+        AssetType assetType = findAssetTypeById(typeId);
+        assetType.setName(type);
+    }
+
+    @Transactional
     public void deleteAssetTypeById(Long id) {
-        assetTypeRepository.deleteById(id);
+        if (assetTypeRepository.findAllById(id).isEmpty()) {
+            assetTypeRepository.deleteById(id);
+            return;
+        }
+        throw new DomainException(ASSET_TYPE_DELETE_ERROR);
     }
 
     public AssetType findAssetTypeById(Long id) {
         return assetTypeRepository.findById(id)
+            .orElseThrow(() -> new DomainException(ASSET_TYPE_NOT_FOUND));
+    }
+
+    public AssetType findAssetTypeByName(String name) {
+        return assetTypeRepository.findByName(name)
             .orElseThrow(() -> new DomainException(ASSET_TYPE_NOT_FOUND));
     }
 
@@ -42,11 +64,15 @@ public class AssetTypeService {
             .orElseThrow(() -> new DomainException(ASSET_TYPE_NOT_FOUND));
     }
 
-    public List<AssetType> findAllParentAssetType() {
-        return assetTypeRepository.findAllParentType();
+    public List<AssetTypeParentResult> findAllParentAssetType() {
+        return assetTypeRepository.findAllParentType()
+            .stream().map(AssetTypeParentResult::from)
+            .toList();
     }
 
     public List<AssetType> findAll() {
         return assetTypeRepository.findAll();
     }
+
+
 }
